Made a few more changes to Type system
It actually doesn't make sense to allow literal-encoded enums of types
other than integer, because the byte encoding of the data isn't canonical,
so there wouldn't be any way to really look up the value by its literal
bytes.  This can still work with integers though, and that was the main
use case anyway.