Refactor userp_bstr to be a struct + array
The previous design with array appended to struct was more work
in most of the cases where I used it, so I finally decided to
switch to a more conventional vector design with a struct that
is held in the parent record, or stack.  This makes the 'count'
or 'alloc' always available without a NULL check.