Refactored diag buffers, and implemented scope parsing
The modes where buffers were stored in userp_diag were flaky, because
the env would reference a buffer that was holding a strong reference to
the env, which required workarounds.  Now the USERP_DIAG_BUFSTR copies
the characters into a the struct.  Also, when it prints the string, it
quotes it and escapes control characters and unicode, which should be
more useful for debugging.

Then, I did the big overhaul on the symbol table parsing code.