Fix bugs in symtable, add trace logging, tune allocation, 'pair' optimization.
I fixed the last bugs in the hashtree.  Then I added better logging to see
what was going on in the future by flipping the log_trace flag.  Then with
the logging, I was able to tune the allocations as the tree grows.  Then
I added an optimization where the first hash collision allocates a single
node which references both symbols (a 'pair') and if there is another
collision later it upgrades this node to two nodes before adding the third
to form a real tree.  This roughly cuts the number of extra nodes needed in
half, causing fewer reallocations.