Redesign varqty in terms of 32-bit 'limbs'
This finally gives the decoder for varqty a sensible API, and fairly
good efficiency.  The decoder requires output to be multiples of 4 bytes,
and the protocol no longer has a special case for 60 bits, opting instead
to go direct to bigint, but checking for a 32 and 64 bit special cases of
bigint.

Also added some new utilities in the header for working with endian-ness.

Added more unit tests and all are passing.